Reorganize packages, add some javadoc
