+++ lightblue/src/main/java/org/esbtools/eventhandler/lightblue/NotificationFactory.java
@@ -18,6 +18,7 @@
 
 package org.esbtools.eventhandler.lightblue;
 
+import org.esbtools.eventhandler.lightblue.client.LightblueRequester;
 import org.esbtools.lightbluenotificationhook.NotificationEntity;
 
 public interface NotificationFactory {
