+++ lightblue/src/main/java/org/esbtools/eventhandler/lightblue/client/LightblueErrors.java
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.esbtools.eventhandler.lightblue;
+package org.esbtools.eventhandler.lightblue.client;
 
 import com.redhat.lightblue.client.model.DataError;
 import com.redhat.lightblue.client.model.Error;
