+++ lightblue/src/main/java/org/esbtools/eventhandler/lightblue/client/LightblueRequester.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2015 esbtools Contributors and/or its affiliates.
+ *  Copyright 2016 esbtools Contributors and/or its affiliates.
  *
  *  This file is part of esbtools.
  *
@@ -16,13 +16,13 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.esbtools.eventhandler.lightblue;
+package org.esbtools.eventhandler.lightblue.client;
 
 import com.redhat.lightblue.client.request.AbstractLightblueDataRequest;
 import com.redhat.lightblue.client.response.LightblueDataResponse;
 
-import org.esbtools.eventhandler.Responses;
+import org.esbtools.eventhandler.Requester;
 
-public interface LightblueResponses extends
-        Responses<AbstractLightblueDataRequest, LightblueDataResponse> {
+public interface LightblueRequester extends
+        Requester<AbstractLightblueDataRequest, LightblueDataResponse> {
 }
