+++ lightblue/src/main/java/org/esbtools/eventhandler/lightblue/config/EventHandlerConfigEntity.java
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.esbtools.eventhandler.lightblue.model;
+package org.esbtools.eventhandler.lightblue.config;
 
 import org.esbtools.eventhandler.lightblue.LightblueDocumentEventRepositoryConfig;
 import org.esbtools.eventhandler.lightblue.LightblueNotificationRepositoryConfig;
