+++ lightblue/src/main/java/org/esbtools/eventhandler/lightblue/config/MutableLightblueDocumentEventRepositoryConfig.java
@@ -16,7 +16,9 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.esbtools.eventhandler.lightblue;
+package org.esbtools.eventhandler.lightblue.config;
+
+import org.esbtools.eventhandler.lightblue.LightblueDocumentEventRepositoryConfig;
 
 import javax.annotation.concurrent.ThreadSafe;
 import java.time.Duration;
