+++ lightblue/src/main/java/org/esbtools/eventhandler/lightblue/locking/LightblueAutoPingLockStrategy.java
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.esbtools.eventhandler.lightblue;
+package org.esbtools.eventhandler.lightblue.locking;
 
 import com.redhat.lightblue.client.LightblueException;
 import com.redhat.lightblue.client.Locking;
