+++ lightblue/src/main/java/org/esbtools/eventhandler/lightblue/locking/LockStrategy.java
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.esbtools.eventhandler.lightblue;
+package org.esbtools.eventhandler.lightblue.locking;
 
 /**
  * Abstracts an atomic exclusive locking mechanism which is used to coordinate concurrent work on
