+++ lightblue/src/main/java/org/esbtools/eventhandler/lightblue/locking/LostLockException.java
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.esbtools.eventhandler.lightblue;
+package org.esbtools.eventhandler.lightblue.locking;
 
 import java.util.Collections;
 import java.util.List;
