+++ lightblue/src/test/java/org/esbtools/eventhandler/lightblue/testing/MultiStringNotification.java
@@ -22,7 +22,7 @@ import com.google.common.base.Joiner;
 import com.google.common.util.concurrent.Futures;
 import org.esbtools.eventhandler.DocumentEvent;
 import org.esbtools.eventhandler.lightblue.LightblueNotification;
-import org.esbtools.eventhandler.lightblue.LightblueRequester;
+import org.esbtools.eventhandler.lightblue.client.LightblueRequester;
 import org.esbtools.lightbluenotificationhook.NotificationEntity;
 
 import java.sql.Date;
